hydroponic_system/requirements.txt
asgiref==3.8.1
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-filter==24.2
django-templated-mail==1.1.1
django-url-filter==0.3.15
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-url-filters==0.5.1
enum-compat==0.0.3
idna==3.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.1
voluptuous==0.14.2
psycopg2